+++ HAPI_Start/Vector3.cpp
@@ -1,38 +1,38 @@
 #include "Vector3.h"
 
-Vector3::Vector3(int x, int y, int z)
+Vector3::Vector3(float x, float y, float z)
 {
 	m_x = x;
 	m_y = y;
 	m_z = z;
 }
 
-void Vector3::SetX(int x)
+void Vector3::SetX(float x)
 {
 	m_x = x;
 }
 
-void Vector3::SetY(int y)
+void Vector3::SetY(float y)
 {
 	m_y = y;
 }
 
-void Vector3::SetZ(int z)
+void Vector3::SetZ(float z)
 {
 	m_z = z;
 }
 
-int Vector3::GetX()
+float Vector3::GetX()
 {
 	return m_x;
 }
 
-int Vector3::GetY()
+float Vector3::GetY()
 {
 	return m_y;
 }
 
-int Vector3::GetZ()
+float Vector3::GetZ()
 {
 	return m_z;
 }
