+++ HAPI_Start/Vector3.h
@@ -2,15 +2,15 @@
 
 struct Vector3 {
 private:
-	int m_x = 0;
-	int m_y = 0;
-	int m_z = 0;
+	float m_x = 0;
+	float m_y = 0;
+	float m_z = 0;
 public:
-	Vector3(int x = 0, int y = 0, int z = 0);
-	void SetX(int x);
-	void SetY(int y);
-	void SetZ(int z);
-	int GetX();
-	int GetY();
-	int GetZ();
+	Vector3(float x = 0, float y = 0, float z = 0);
+	void SetX(float x);
+	void SetY(float y);
+	void SetZ(float z);
+	float GetX();
+	float GetY();
+	float GetZ();
 };
